worker-integration(job-b4ce): mario/Enemy/telesa::init__7TTelesaFP12TLiveManager [checkpoint 778a6432]

+++ src/Enemy/telesa.cpp
@@ -242,7 +242,7 @@ void TTelesa::init(TLiveManager* manager)
 	                     *img);
 	mMActor->setLightType(LIGHT_TYPE_INDIRECT);
 	if (mInstanceIndex == 0) {
-		for (u16 i = 0; i < getModel()->getModelData()->getJointNum(); ++i)
+		for (u8 i = 0; i < getModel()->getModelData()->getJointNum(); ++i)
 			;
 	}
